Lab2: implement all endpoints, refactor, fix bugs

+++ Lab2/Lab2/Services/ContactService.cs
@@ -1,9 +1,9 @@
-using System.Linq.Expressions;
 using AutoMapper;
 using Lab2.Data;
 using Lab2.DTOs.Contact;
 using Lab2.DTOs.QueryParameters;
 using Lab2.Entities;
+using Lab2.Exceptions;
 using Lab2.Repositories.Interfaces;
 using Lab2.Services.Interfaces;
 
@@ -21,29 +21,62 @@ public class ContactService : IContactService
         _mapper = mapper;
         _unitOfWork = unitOfWork;
     }
+    
+    private async Task ValidateSavingContact(ContactDto contactDto)
+    {
+        // For update operation, check if any other account with this phone or email exists
+        // For add operation, dto id will be 0, so we can also use this method
+        
+        //  Check phone field uniqueness if it modified and not empty
+        if (!string.IsNullOrEmpty(contactDto.Phone))
+        {
+           if (!await _contactRepository.IsExistAsync(a => a.Phone == contactDto.Phone 
+                                                           && a.Id != contactDto.Id))
+               throw new EntityValidationException("This phone number is already taken"); 
+        } 
+        
+        // Check email field uniqueness if it modified and not empty
+        if (!string.IsNullOrEmpty(contactDto.Email))
+        {
+           if (!await _contactRepository.IsExistAsync(a => a.Email == contactDto.Email 
+                                                           && a.Id != contactDto.Id))
+               throw new EntityValidationException("This email is already taken"); 
+        }
+    }
 
     public async Task<ContactDto> CreateAsync(ContactDto contactDto)
     {
+        // Check phone and email uniqueness
+        await ValidateSavingContact(contactDto);
+        
         var contact = _mapper.Map<Contact>(contactDto);
         _contactRepository.Add(contact);
         await _unitOfWork.CommitAsync();
         return _mapper.Map<ContactDto>(contact);
     }
 
-    public async Task<bool> DeleteAsync(int contactId)
+    public async Task DeleteAsync(int contactId)
     {
+        // Check if contact is exist
+        if (!await _contactRepository.IsExistAsync(c => c.Id == contactId))
+            throw new EntityNotFoundException($"Contact with id {contactId} is not found");
+        
         _contactRepository.Delete(new Contact { Id = contactId });
-        return await _unitOfWork.CommitAsync() > 0;
+        await _unitOfWork.CommitAsync();
     }
 
-    public async Task<ContactDto> GetByIdAsync(int id)
+    public async Task<ContactDto?> GetByIdAsync(int id)
     {
         return _mapper.Map<ContactDto>(await _contactRepository.GetByIdAsync(id));
     }
 
     public async Task<PagedResult<ContactDto>> GetListAsync(ContactQueryParameters cqp)
     {
-        var (contacts, contactCount) = await _contactRepository.GetContactPagedListAsync(cqp);
+        var (contacts, contactCount) = await _contactRepository.GetContactPagedListAsync(search: cqp.Search,
+                                                                                         orderBy: cqp.OrderBy,
+                                                                                         skip: (cqp.PageIndex - 1) * cqp.PageSize,
+                                                                                         take: cqp.PageSize,
+                                                                                         isDescending: cqp.IsDescending);
         var result = _mapper.Map<List<ContactDto>>(contacts);
 
         return new PagedResult<ContactDto>(result, contactCount, cqp.PageIndex, cqp.PageSize);
@@ -51,7 +84,13 @@ public class ContactService : IContactService
 
     public async Task<ContactDto> UpdateAsync(ContactDto contactDto)
     {
+        // Check phone and email uniqueness
+        await ValidateSavingContact(contactDto);
+        
         var contact = await _contactRepository.GetByIdAsync(contactDto.Id);
+        if (contact == null)
+            throw new EntityNotFoundException($"Contact with id {contactDto.Id} is not found");
+        
         _mapper.Map(contactDto, contact);
         await _unitOfWork.CommitAsync();
         return _mapper.Map<ContactDto>(contact);
